feat(editor): add comparators for gradual updating of Interactive components
This is a start on getting "live updates" into the controls. As code is editor,
we monitor the structure via the new Source classes and are able to diff and
update resources, using a system backed by fast-json-patch which we already use
for incremental history.

This is not functional with this PR, but is set to be wired in soon. I'll
likely take a break on this stuff for the next couple days to work on service
implementation but will swing back around here by the following week.