feat(*): start the process of getting Interactive connections going to prod
This is wide-ranging commit since a lot of different components have to work to
get any real "result" going. The end goal now is getting a bundle working E2E,
and as an intermediate step have the editor be able to connect to a channel
in prod today. To that end, these are the general changes:

 - Added a `pack` command to create a bundle as it will be uploaded to our serves
 - Fleshed out the dev server API to add login and connect endpoints. These
   proxy to the production endpoints for the most part, but that can be
   overridden to point at standalone
 - Added the UI for logging in and connecting in the editor.
 - WIP: a Participant class that the editor and our production frontend will use
   to actually connect to Interactive.
 - WIP: state changes in the editor to handle being connected. The idea is that
   the code editor on the side will be disabled for input but update with the
   current control state tree displayed on the RHS.

Some screen grabs posted in Slack earlier:

 - https://peet.io/i/17-08-ctl9l52qhnb2j7mql2e5zat44iik3x.mp4
 - https://peet.io/i/17-08-ptgckpcxtzxzwzmjqgprl7kg45hkxl.mp4