feat(*): get controls working end to end, live state display and editing
Controls now connect and run E2E against a standalone tetrisd instance using
unmodified pre-CC controls and scripts. When connected in production, state
is displayed in the code editor and updated in realtime. This is done through
the Participant class, which I intend to use in the production frontend for
connections as well. I did some extra work (that is, outside of functionality
that prod needs) to make sure it can gracefully handle webpack's hot reloads
without interruption.

I moved some of the profile around so that multiple sets of credentials can be
stored for different hosts. This won't be needed for most users but is useful
for talking to standalone and could be useful if first parties have custom
instances running anywhere. By default we'll always connect to the user's own
channel, but I hid a dialog behind a shift+click of the rocket to change that;
we don't really want people connecting to random channels and potentially
throwing bad input to game clients, but it's quite useful for our own testing
and anyone who knows code and has any amount of determination could do it
even without that.

Some screenies:

 - https://peet.io/i/17-08-l8tdkgh07e2h43rl4p78hz979lnbq5.mp4 channel selection dialog, some status messages
 - https://peet.io/i/17-08-j5pz1123sdff1c36ql1vyvsswivde7.mp4 hot reloading