feat(editor): prompt to rename on conflict, general better error structure
Previously, our HTTP errors were a bit willy-nilly. This introduces some
structure to them with a corresponding Angular service to catch them somewhat
nicely.

It adds a dialog which is popped open if a user tries to upload on a bundle
they don't have access to, allowing them to rename it in the UI instead of
throwing a cryptic 403.