fix: not show layout screen if HTML template selected
This change fixes the new project wizard so the user will only see the template page if they have chosen the preact template. This is because the layout choice does not apply for the HTML template.

This change also adds simple back stack logic because the wizard now has a branching flow.