feat(emulation): allow full range of css dimensions to set video position (fixes #112)
This refactors some of the logic we had before and also fixes some bugs there.
Essentially, rather than figuring out what the 16:9 ratio we need to set the
video is based on the position directly, we apply the position and then call
getClientBoundingRect to determine the outer dimensions, then size the video
accordingly and send it back down to the controls. This is generally more
robust.